+++ es/lib/isHash.js
@@ -16,6 +16,6 @@ var lengths = {
 };
 export default function isHash(str, algorithm) {
   assertString(str);
-  var hash = new RegExp("^[a-f0-9]{".concat(lengths[algorithm], "}$"));
+  var hash = new RegExp("^[a-fA-F0-9]{".concat(lengths[algorithm], "}$"));
   return hash.test(str);
 }