+++ es/lib/isHexadecimal.js
@@ -1,5 +1,5 @@
 import assertString from './util/assertString';
-var hexadecimal = /^[0-9A-F]+$/i;
+var hexadecimal = /^(0x|0h)?[0-9A-F]+$/i;
 export default function isHexadecimal(str) {
   assertString(str);
   return hexadecimal.test(str);